Move ContentHeading to folder 'heading'
