+++ src/components/heading/index.ts
