+++ src/stacks-hierarchy/Root_PurchaseOverviewScreen/components/FlexTicketDiscountInfo.tsx
@@ -21,7 +21,7 @@ import {
   LinkSectionItem,
   Section,
 } from '@atb/components/sections';
-import {ContentHeading} from '@atb/components/content-heading';
+import {ContentHeading} from '@atb/components/heading';
 
 type Props = {
   userProfiles: UserProfileWithCountAndOffer[];
